# server/static/swagger.yaml
openapi: 3.0.3
info:
  title: Vietnamese Text Summarizer
  description: 'This is a guide for using API by TinDaKing'
  version: 1.0.11
servers:
  - url: http://localhost:5000

paths:
  
  /summarize:
    post:
      tags:
        - Summarize
      summary: Sumarize input text or paragraph
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Summarize'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummarizeResult'
        '400':
          description: No JSON data received
        '500':
          description: Something wrong while running the model
          
  /register:
    post:
      tags:
        - User
      summary: Create new account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: User created successfully
        '400':
          description: No JSON data received
        '500':
          description: Error while accessing database
                
  /login:
    post:
      tags:
        - User
      summary: Log in existing account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Logged in successfully
        '400':
          description: No JSON data received
        '401':
          description: Wrong username or password
        '500':
          description: Error while accessing database
    get:
      tags:
        - User
      summary: Try to log in while already logged in
      description: 'Redirect to home page'
      responses:
        '200':
          description: Logged in successfully
          
  /logout:
    get:
      tags:
        - User
      summary: Log out user from the session
      description: Log out user from the session
      
      responses:
        '401':
          description: Successful log out
          
  /resetpassword:
    post:
      tags:
        - User
      summary: Reset password
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPassword'
      responses:
        '200':
          description: Logged in successfully
        '400':
          description: No JSON data received
        '401':
          description: Wrong username or password
        '500':
          description: Error while accessing database
          

  /home:
    get:
      tags:
        - Home
      summary: Go to main home page after log in
      
      responses:
        '200':
          description: Successful operation
        '401':
          description: User have not logged in
    

components:
  schemas:
    Summarize:
      type: object
      properties:
        input-text:
          type: string
          description: input text
          example: Triển lãm di động MWC 2024, diễn ra ngày 26-29/2 tại Barcelona, quy tụ nhiều sản phẩm cho thấy tương lai của công nghệ, như điện thoại cong, nhẫn thông minh.Tại MWC năm nay, có hai hãng lớn cùng giới thiệu smatrtphone có thể cuộn vào cổ tay là Motorola và Samsung. Trong đó, nguyên mẫu của Motorola từng trưng bày tại sự kiện Lenovo Tech World hồi tháng 10/2023, còn thiết bị Samsung lần đầu được giới thiệu. Nguyên mẫu điện thoại Motorola có thiết kế dạng thanh, màn hình POLED uốn cong với kích thước 6,9 inch, độ phân giải Full HD. Phía sau máy làm bằng chất liệu nhựa, bọc vải và kết cấu dạng khớp để uốn phủ lên cổ tay. Còn Samsung mang tới nguyên mẫu Cling Band với thiết kế tương tự nhưng các tính năng hạn chế hơn thiết bị từ Motorola, theo CNet. Hãng điện tử Hàn Quốc chưa hé lộ kế hoạch thương mại hóa thiết bị đeo này. Nguyên mẫu smartphone dùng công nghệ E-Color Shift của Infinix được trang bị màn hình e-ink phía sau. Khác với mẫu YotaPhone trước đây chỉ có hai màu đen trắng, thiết bị mới sử dụng màn hình màu, cũng như lưu giữ được hình ảnh tĩnh ngay cả khi không có nguồn điện.Theo Infinix, việc sử dụng màn hình e-ink kết hợp với công nghệ riêng giúp bề mặt sau của máy có thể thay đổi nhưng tiêu hao năng lượng không đáng kể. Bằng cách áp dụng các điện áp khác nhau, điện trường bên trong vi cấu trúc sẽ thay đổi, khiến các hạt màu tương ứng di chuyển và hiển thị màu sắc theo mong muốn mà không tiêu tốn điện năng, đại diện Infinix nói với TechCrunch.
    SummarizeResult:
      type: object
      properties:
        message:
          type: string
          example: Input text received successfully
        output-text:
          type: string
          example: Triển lãm di động MWC 2024, diễn ra ngày 26-29/2 tại Barcelona, quy tụ nhiều sản phẩm cho thấy tương lai của công nghệ, như điện thoại cong, nhẫn thông minh.Tại MWC năm nay, có hai hãng lớn cùng giới thiệu smatrtphone có thể cuộn vào cổ tay là Motorola và Samsung. Trong đó, nguyên mẫu của Motorola từng trưng bày tại sự kiện Lenovo Tech World hồi tháng 10/2023, còn thiết bị Samsung lần đầu được giới thiệu. Nguyên mẫu điện thoại Motorola có thiết kế dạng thanh, màn hình POLED uốn cong với kích thước 6,9 inch, độ phân giải Full HD. Còn Samsung mang tới nguyên mẫu Cling Band với thiết kế tương tự nhưng các tính năng hạn chế hơn thiết bị từ Motorola, theo CNet. Hãng điện tử Hàn Quốc chưa hé lộ kế hoạch thương mại hóa thiết bị đeo này.
        score:
          type: number
          example: 0.8022163566921933
          
    UserRegister:
      type: object
      properties:
        username:
          type: string
          example: tranhoangtin
        password:
          type: string
          example: 1234
          
    UserLogin:
      type: object
      properties:
        username:
          type: string
          example: tranhoangtin
        password:
          type: string
          example: 1234
          
    ResetPassword:
      type: object
      properties:
        username:
          type: string
          example: tranhoangtin
        password:
          type: string
          example: Hoangtin1234
    
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header